requirements.txt
python-dotenv
streamlit
google-generativeai
